Removed USB handling from interrupt context.
The buffer size of 8 pulses for read and write
are too low with the current interrupt load.
As it was never really required to have USB handled
via IRQs, I've decided to remove it as USB is not
as time sensitive as the floppy signal handling.

One other option would have been to activate nested interrupts.
Maybe for the future.