Experimental Pasti/stx and NFRA support
This format is truly horrible as it doesn't contains
the actual data stored on the disk but instead what the
floppy disk controller has interpreted by reading it.
Therefore the stx support will always be flaky.

For the Turrican.stx protection it is now required
to support non flux reversal areas (NFRA).
This is the biggest change here for the firmware
and I'm not entirely happy with it as it is not reliable.