Tool set for regression testing
Added unit test which reduces an analyzed
image and its track data to a single hash value.
This allows regression testing on the data which will
be sent to the USB device.
It is not as reliable as actually writing the disk
but it is much faster and can be done without any hardware.

Added an option to the tool which prints out
the raw cell data as hexdump.
This allows a more in depth search for differences
in case a hash value is not as expected.

Also cleaned up main.rs and moved some code
to separate files.