G64 maintenance and weak bits
Added weak bits.
Added lots of images to the list.
Added command line option to select tracks for writing.

Non Flux Reversal Area and Weak Bits
look the same in the data stream and a configuration
flag was added to communicate what has to be done
to the device.

This however means that NFRA and Weak Bits can't occur
on the same track!